authMW update

+++ Controllers/eventController.js
@@ -13,10 +13,10 @@ module.exports.getAllEvents=(request,response)=>{
     
 }
 module.exports.getEventById=(request,response)=>{
-    // if(request.role !=="admin" || request.role !=="student")
-    // {
-    //    throw new Error("Not Authorizd");
-    // }
+    if(request.role !=="admin" && request.role !=="student" && request.role!=="speaker")
+    {
+       throw new Error("Not Authorizd");
+    }
     Event.findById({_id:request.params.id})
            .then(data=>{
                response.status(200).json(data);
@@ -25,18 +25,18 @@ module.exports.getEventById=(request,response)=>{
 }
 
 module.exports.createEvent=(request,response,next)=>{
-    // if(request.role !=="admin")
-    // {
-    //    throw new Error("Not Authorizd");
-    // }
-    // let result = validationResult(request);
-    // if(!result.isEmpty()){
-    //     let message=result.array().reduce((current,error)=>current+error.msg," ");
-    //     let error = new Error(message);
-    //     error.status=422;
-    //     throw error;
+    if(request.role !=="admin")
+    {
+       throw new Error("Not Authorizd");
+    }
+    let result = validationResult(request);
+    if(!result.isEmpty()){
+        let message=result.array().reduce((current,error)=>current+error.msg," ");
+        let error = new Error(message);
+        error.status=422;
+        throw error;
 
-    // }
+    }
     let event = new Event({
         _id:request.body._id,
         title:request.body.title,
@@ -55,10 +55,10 @@ module.exports.createEvent=(request,response,next)=>{
 
 
 module.exports.updateEvent=(request,response,next)=>{
-    // if(request.role !=="admin")
-    // {
-    //    throw new Error("Not Authorizd");
-    // }
+    if(request.role !=="admin")
+    {
+       throw new Error("Not Authorizd");
+    }
     Event.updateOne({_id:request.params.id},{
         $set:{
          
@@ -82,10 +82,10 @@ module.exports.updateEvent=(request,response,next)=>{
 
 module.exports.deleteEvent=(request,response,next)=>{
 
-        // if(request.role !=="admin")
-        // {
-        //     throw new Error("Not Authorizd");
-        // }
+        if(request.role !=="admin")
+        {
+            throw new Error("Not Authorizd");
+        }
    
         Event.deleteOne({_id:request.params.id},{
