+++ Controllers/speakerController.js
@@ -2,6 +2,7 @@ const {validationResult}=require("express-validator");
 const Speaker=require("../Models/speakerModel");
 
 module.exports.getAllSpeakers=(request,response)=>{
+
     Speaker.find({})
            .then((data)=>{
                response.status(200).json(data);
@@ -12,10 +13,10 @@ module.exports.getAllSpeakers=(request,response)=>{
     
 }
 module.exports.getSpeakerById=(request,response)=>{
-    // if(request.role !=="admin" || request.role !=="student")
-    // {
-    //    throw new Error("Not Authorizd");
-    // }
+    if(request.role !=="admin" && request.role !=="speaker")
+    {
+       throw new Error("Not Authorizd");
+    }
     Speaker.findById({_id:request.params.id})
            .then(data=>{
                response.status(200).json(data);
@@ -25,10 +26,10 @@ module.exports.getSpeakerById=(request,response)=>{
 
 module.exports.createSpeaker=(request,response,next)=>{
 
-    // if(request.role !=="admin")
-    // {
-    //    throw new Error("Not Authorizd");
-    // }
+    if(request.role !=="admin" && request.role !=="speaker")
+    {
+       throw new Error("Not Authorizd");
+    }
     let result = validationResult(request);
     if(!result.isEmpty()){
         let message=result.array().reduce((current,error)=>current+error.msg," ");
@@ -44,7 +45,8 @@ module.exports.createSpeaker=(request,response,next)=>{
         password:request.body.password, 
         city: request.body.city, 
         street:request.body.street, 
-        building:request.body.building
+        building:request.body.building,
+        role:request.body.role
     })
     speaker.save()
     .then((data)=>{ 
@@ -57,10 +59,10 @@ module.exports.createSpeaker=(request,response,next)=>{
 
 module.exports.updateSpeaker=(request,response,next)=>{
 
-    // if(request.role !=="admin" && request.role!=="speaker")
-    // {
-    //    throw new Error("Not Authorizd");
-    // }
+    if(request.role !=="admin" && request.role !=="speaker")
+    {
+       throw new Error("Not Authorizd");
+    }
     Speaker.updateOne({_id:request.params.id},{
         $set:{
             email:request.body.email,
@@ -83,10 +85,10 @@ module.exports.updateSpeaker=(request,response,next)=>{
 
 
 module.exports.deleteSpeaker=(request,response,next)=>{
-    //    if(request.role !=="admin")
-    //    {
-    //         throw new Error("Not Authorizd");
-    //    }  
+       if(request.role !=="admin")
+       {
+            throw new Error("Not Authorizd");
+       }  
         Speaker.deleteOne({_id:request.params.id},{
           
         }).then(data=>{
